requirements.txt
web3==5.25.0
pandas==1.3.5
numpy==1.21.4
matplotlib==3.5.0
nltk==3.6.7
python-dotenv==0.19.2
